Several adjustments to use blurhash, document type thumbnail and srcset on gutenberg blocks. #456 #467 #463
