+++ src/assets/css/tainacan-gutenberg-block-carousel-collections-list.css
@@ -18,16 +18,16 @@
         display: flex;
         display: -ms-grid;
         display: grid;
-        -ms-grid-columns: 33% 33% 33%;
-        grid-template-columns: 33% 33% 33%;
-        -ms-grid-rows: 50% 50%;
-        grid-template-rows: 50% 50%;
-        grid-column-gap: 2px;
-        grid-row-gap: 2px;
+        -ms-grid-columns: 1fr 1fr 1fr;
+        grid-template-columns: 1fr 1fr 1fr;
+        -ms-grid-rows: 1fr 1fr;
+        grid-template-rows: 1fr 1fr;
+        grid-gap: 5px;
+        gap: 5px;
         width: 60px;
         height: 40px;
         margin: 4px auto 8px auto; }
-        .components-panel__body .collection-carousel-view-modes button.collection-carousel-view-mode-grid div > div:first-of-type {
+        .components-panel__body .collection-carousel-view-modes button.collection-carousel-view-mode-grid div > *:first-of-type {
           flex-basis: 100%;
           -ms-grid-column: 1;
           -ms-grid-column-span: 2;
@@ -35,12 +35,12 @@
           -ms-grid-row-span: 2;
           grid-column: 1/3;
           grid-row: 1/3;
-          width: 41px;
-          height: 42px; }
-        .components-panel__body .collection-carousel-view-modes button.collection-carousel-view-mode-grid div > div {
+          width: 100%;
+          height: 100%; }
+        .components-panel__body .collection-carousel-view-modes button.collection-carousel-view-mode-grid div > * {
           flex-basis: 50%;
-          width: 20px;
-          height: 20px;
+          width: 100%;
+          height: 100%;
           background: var(--tainacan-block-gray4, #555758);
           transition: background-color ease 0.5s; }
     .components-panel__body .collection-carousel-view-modes button.collection-carousel-view-mode-thumbnail {
@@ -139,11 +139,13 @@
         display: flex;
         display: -ms-grid;
         display: grid;
-        grid-template-columns: 33% 33% 33%;
-        grid-template-rows: 50% 50%;
+        grid-template-columns: 1fr 1fr 1fr;
+        grid-template-rows: 1fr 1fr;
         width: 100%;
+        grid-gap: 5px;
+        gap: 5px;
         box-sizing: border-box; }
-        .wp-block-tainacan-carousel-collections-list .tainacan-carousel .swiper-container .swiper-slide.collection-list-item-grid .collection-items-grid img:first-of-type {
+        .wp-block-tainacan-carousel-collections-list .tainacan-carousel .swiper-container .swiper-slide.collection-list-item-grid .collection-items-grid > *:first-of-type {
           flex-basis: 100%;
           -ms-grid-column: 1;
           -ms-grid-column-span: 2;
@@ -151,11 +153,10 @@
           -ms-grid-row-span: 2;
           grid-column: 1/3;
           grid-row: 1/3; }
-        .wp-block-tainacan-carousel-collections-list .tainacan-carousel .swiper-container .swiper-slide.collection-list-item-grid .collection-items-grid img {
+        .wp-block-tainacan-carousel-collections-list .tainacan-carousel .swiper-container .swiper-slide.collection-list-item-grid .collection-items-grid > * {
           flex-basis: 50%;
           width: 100%;
           height: auto;
-          padding: 2px;
           margin-bottom: 0px; }
   .wp-block-tainacan-carousel-collections-list .preview-warning {
     width: 100%;
@@ -313,9 +314,11 @@
         display: flex;
         display: -ms-grid;
         display: grid;
-        grid-template-columns: 33% 33% 33%;
-        grid-template-rows: 50% 50%; }
-        .wp-block-tainacan-carousel-collections-list ul.collections-list-edit li.collection-list-item .collection-items-grid img:first-of-type {
+        grid-template-columns: 1fr 1fr 1fr;
+        grid-template-rows: 1fr 1fr;
+        grid-gap: 5px;
+        gap: 5px; }
+        .wp-block-tainacan-carousel-collections-list ul.collections-list-edit li.collection-list-item .collection-items-grid > *:first-of-type {
           flex-basis: 100%;
           -ms-grid-column: 1;
           -ms-grid-column-span: 2;
@@ -323,9 +326,8 @@
           -ms-grid-row-span: 2;
           grid-column: 1/3;
           grid-row: 1/3; }
-        .wp-block-tainacan-carousel-collections-list ul.collections-list-edit li.collection-list-item .collection-items-grid img {
+        .wp-block-tainacan-carousel-collections-list ul.collections-list-edit li.collection-list-item .collection-items-grid > * {
           flex-basis: 50%;
-          padding: 2px;
           margin-bottom: 0px; }
   @media only screen and (max-width: 1686px) {
     .wp-block-tainacan-carousel-collections-list ul.collections-list-edit li.collection-list-item {
