+++ src/views/gutenberg-blocks/tainacan-collections/carousel-collections-list/carousel-collections-list.scss
@@ -19,27 +19,27 @@
             margin-right: 6px;
             div {
                 @include display-grid;
-                -ms-grid-columns: 33% 33% 33%;
-                grid-template-columns: 33% 33% 33%;
-                -ms-grid-rows: 50% 50%;
-                grid-template-rows: 50% 50%;
-                grid-column-gap: 2px;
-                grid-row-gap: 2px;
+                -ms-grid-columns: 1fr 1fr 1fr;
+                grid-template-columns: 1fr 1fr 1fr;
+                -ms-grid-rows: 1fr 1fr;
+                grid-template-rows: 1fr 1fr;
+                grid-gap: 5px;
+                gap: 5px;
                 width: 60px;
                 height: 40px;
                 margin: 4px auto 8px auto;
 
-                &>div:first-of-type {
+                &>*:first-of-type {
                     flex-basis: 100%;
                     @include grid-child(1, 3, 1, 3);
-                    width: 41px;
-                    height: 42px;
+                    width: 100%;
+                    height: 100%;
                 }
 
-                &>div {
+                &>* {
                     flex-basis: 50%; 
-                    width: 20px;
-                    height: 20px;
+                    width: 100%;
+                    height: 100%;
                     background: var(--tainacan-block-gray4, $gray4);
                     transition: background-color ease 0.5s; 
                 }
@@ -163,21 +163,22 @@
                 }
                 .collection-items-grid {
                     @include display-grid;
-                    grid-template-columns: 33% 33% 33%;
-                    grid-template-rows: 50% 50%;
+                    grid-template-columns: 1fr 1fr 1fr;
+                    grid-template-rows: 1fr 1fr;
                     width: 100%;
+                    grid-gap: 5px;
+                    gap: 5px;
                     box-sizing: border-box;
     
-                    img:first-of-type {
+                    &>*:first-of-type {
                         flex-basis: 100%;
                         @include grid-child(1, 3, 1, 3);
                     }
     
-                    img {
+                    &>* {
                         flex-basis: 50%; 
                         width: 100%;
                         height: auto;
-                        padding: 2px;
                         margin-bottom: 0px; 
                     }
                 }    
@@ -378,17 +379,18 @@
 
             .collection-items-grid {
                 @include display-grid;
-                grid-template-columns: 33% 33% 33%;
-                grid-template-rows: 50% 50%;
+                grid-template-columns: 1fr 1fr 1fr;
+                grid-template-rows: 1fr 1fr;
+                grid-gap: 5px;
+                gap: 5px;
 
-                img:first-of-type {
+                &>*:first-of-type {
                     flex-basis: 100%;
                     @include grid-child(1, 3, 1, 3);
                 }
 
-                img {
+                &>* {
                     flex-basis: 50%; 
-                    padding: 2px;
                     margin-bottom: 0px; 
                 }
             }
