+++ src/views/gutenberg-blocks/tainacan-items/carousel-items-list/carousel-items-list-theme.js
@@ -1,5 +1,7 @@
 import Vue from 'vue';
 import CarouselItemsListTheme from './carousel-items-list-theme.vue';
+import { ThumbnailHelperPlugin } from '../../../admin/js/utilities.js';
+import VueBlurHash from 'vue-blurhash';
 
 // Vue Dev Tools!
 Vue.config.devtools = process && process.env && process.env.NODE_ENV === 'development';
@@ -96,6 +98,8 @@ document.addEventListener("DOMContentLoaded", () => {
                 }
             };
 
+            Vue.use(ThumbnailHelperPlugin);
+            Vue.use(VueBlurHash);
             new Vue( Object.assign({ el: '#' + blockId }, vueOptions) );
         }
     }
